Added auto paths for CORI (to Jaci's as waypoints)
Updated gitignore to ignore .idea files.Also added RRRight, LLLeft, RMiddle,
and LMiddle as waypoints to the JaciPathfinding command [nothing tested or
tuned]. TODO: Add command groups to call on the command, as well as save the
paths as files to read from (reduces computation time). Also need to finish
the if-statement that determines which waypoints to use.